+++ precept/__init__.py
@@ -6,3 +6,4 @@ from ._immutable import *  # noqa: F401, F403
 from ._configs import (  # noqa: F401
     ConfigProperty, Config, ConfigFormat, Nestable, config_factory
 )
+from ._executor import AsyncExecutor  # noqa: F401
